# Conflicts:
#	src/lib/axis/motor/servo/tmc/tmc2209/Tmc2209.h
#	src/lib/axis/motor/servo/tmc2209/Tmc2209.h
#	src/lib/axis/motor/stepDir/tmc/legacy/TmcSPI.cpp
#	src/lib/axis/motor/stepDir/tmc/legacy/TmcSPI.h
#	src/lib/axis/motor/stepDir/tmc/legacy/tmc2209/Tmc2209.h
#	src/lib/axis/motor/stepDir/tmcLegacy/LegacyUART.h
#	src/lib/axis/motor/stepDir/tmcLegacy/TmcSPI.cpp
#	src/lib/axis/motor/stepDir/tmcLegacy/TmcSPI.h
#	src/src/HAL/HAL.cpp
#	src/src/HAL/HAL_ANALOG.h
#	src/src/HAL/HAL_FAST_TICKS.h
#	src/src/HAL/esp/ESP32Libraries1.h
#	src/src/HAL/mbed/Rpi2040.h
#	src/src/HAL/mbed/Rpi2350.h
#	src/src/lib/Macros.h
#	src/src/lib/analog/AN_ESP32.cpp
#	src/src/lib/analog/AN_ESP32.h
#	src/src/lib/analog/Analog.cpp
#	src/src/lib/analog/Analog.h
#	src/src/lib/axis/motor/oDrive/ODrive.ino
#	src/src/lib/axis/motor/oDrive/ODriveCanPlus.cpp
#	src/src/lib/axis/motor/oDrive/ODriveCanPlus.h
#	src/src/lib/axis/motor/oDrive/ODriveNew.cpp
#	src/src/lib/axis/motor/oDrive/ODriveNew.h
#	src/src/lib/axis/motor/servo/Servo.ino
#	src/src/lib/axis/motor/servo/ServoCalibrate.cpp
#	src/src/lib/axis/motor/servo/dc/Dc.cpp
#	src/src/lib/axis/motor/servo/dc/Dc.h
#	src/src/lib/axis/motor/servo/dc/DcServoDriver.cpp
#	src/src/lib/axis/motor/servo/dc/DcServoDriver.h
#	src/src/lib/axis/motor/servo/dc/calibration/TrackingVelocity.cpp
#	src/src/lib/axis/motor/servo/dc/calibration/TrackingVelocity.h
#	src/src/lib/axis/motor/servo/dc/eE/EE.cpp
#	src/src/lib/axis/motor/servo/dc/eE/EE.h
#	src/src/lib/axis/motor/servo/dc/pE/PE.cpp
#	src/src/lib/axis/motor/servo/dc/pE/PE.h
#	src/src/lib/axis/motor/servo/dc/tmc/tmc2130/Tmc2130.cpp
#	src/src/lib/axis/motor/servo/dc/tmc/tmc2130/Tmc2130.h
#	src/src/lib/axis/motor/servo/dc/tmc/tmc5160/Tmc5160.cpp
#	src/src/lib/axis/motor/servo/dc/tmc/tmc5160/Tmc5160.h
#	src/src/lib/axis/motor/servo/dcTmcSPI/DcTmcSPI.cpp
#	src/src/lib/axis/motor/servo/dcTmcSPI/DcTmcSPI.h
#	src/src/lib/axis/motor/servo/tmc2209/Tmc2209.cpp
#	src/src/lib/axis/motor/servo/tmc2209/Tmc2209.h
#	src/src/lib/axis/motor/servo/tmc5160/Tmc5160.cpp
#	src/src/lib/axis/motor/servo/tmc5160/Tmc5160.h
#	src/src/lib/axis/motor/stepDir/tmcLegacy/LegacySPI.cpp
#	src/src/lib/axis/motor/stepDir/tmcLegacy/LegacySPI.h
#	src/src/lib/axis/motor/stepDir/tmcLegacy/LegacyUART.cpp
#	src/src/lib/axis/motor/stepDir/tmcLegacy/LegacyUART.h
#	src/src/lib/axis/motor/stepDir/tmcLegacy/TmcSPI.cpp
#	src/src/lib/axis/motor/stepDir/tmcLegacy/TmcSPI.h
#	src/src/lib/axis/motor/stepDir/tmcStepper/StepperSPI.cpp
#	src/src/lib/axis/motor/stepDir/tmcStepper/StepperSPI.h
#	src/src/lib/axis/motor/stepDir/tmcStepper/StepperUART.cpp
#	src/src/lib/axis/motor/stepDir/tmcStepper/StepperUART.h
#	src/src/lib/encoder/bissc/Asc85.cpp
#	src/src/lib/encoder/bissc/Asc85.h
#	src/src/lib/lib.ino
#	src/src/lib/nv/24xx/24XX.cpp
#	src/src/lib/nv/24xx/24XX.h
#	src/src/lib/nv/Nv.cpp
#	src/src/lib/nv/NvBase.cpp
#	src/src/lib/nv/NvBase.h
#	src/src/lib/nv/NvConfig.h
#	src/src/lib/nv/NvIvPartition.cpp
#	src/src/lib/nv/NvIvPartition.h
#	src/src/lib/nv/NvKvPartition16.cpp
#	src/src/lib/nv/NvKvPartition16.h
#	src/src/lib/nv/NvKvPartition32.cpp
#	src/src/lib/nv/NvKvPartition32.h
#	src/src/lib/nv/NvVolume.cpp
#	src/src/lib/nv/NvVolume.h
#	src/src/lib/nv/device/24xxI2C.h
#	src/src/lib/nv/device/DeviceNull.h
#	src/src/lib/nv/device/EepromArduino.h
#	src/src/lib/nv/device/EepromEmuEsp.h
#	src/src/lib/nv/device/EepromEmuM0.h
#	src/src/lib/nv/device/Mb85rcI2C.h
#	src/src/lib/nv/device/NvDeviceBase.h
#	src/src/lib/nv/device/ShimCached.h
#	src/src/lib/nv/device/ShimDelayedCommit.h
#	src/src/lib/nv/eeprom/EEPROM.cpp
#	src/src/lib/nv/eeprom/EEPROM.h
#	src/src/lib/nv/esp/Esp.cpp
#	src/src/lib/nv/esp/Esp.h
#	src/src/lib/nv/m0/M0.cpp
#	src/src/lib/nv/m0/M0.h
#	src/src/lib/nv/mb85rc/MB85RC.cpp
#	src/src/lib/nv/mb85rc/MB85RC.h
#	src/src/lib/serial/Serial_IP_Ethernet_Client.cpp
#	src/src/lib/serial/Serial_IP_Ethernet_Client.h
#	src/src/lib/serial/Serial_ST4_Master.cpp
#	src/src/pinmaps/Pins.Defaults.h
#	src/src/pinmaps/Pins.Esp8266.h
#	src/src/pinmaps/Pins.M0.h
#	src/src/userInterface/UserInterface.cpp
#	src/userInterface/menus/MenuConnect.cpp

+++ src/src/lib/gpioEx/ssr74HC595/Ssr74HC595.cpp
@@ -5,8 +5,6 @@
 
 #if defined(GPIO_DEVICE) && GPIO_DEVICE == SSR74HC595
 
-#include "../tasks/OnTask.h"
-
 #if !defined(GPIO_SSR74HC595_LATCH_PIN) || GPIO_SSR74HC595_LATCH_PIN == OFF
   #error "GPIO device SSR74HC595 GPIO_SSR74HC595_LATCH_PIN must be present and not OFF."
 #endif
@@ -42,7 +40,7 @@
 #endif
 
 // designed for a 20MHz max bit rate
-IRAM_ATTR void shiftOut20MHz(uint32_t val) {
+IRAM_ATTR void shiftOut20MHz(uint8_t val) {
   if ((val & 0b10000000) == 0) { GPIO_SSR74HC595_DATA_LOW(); } else { GPIO_SSR74HC595_DATA_HIGH(); }
   GPIO_SSR74HC595_CLOCK_HIGH();
   GPIO_SSR74HC595_CLOCK_LOW();
